Added neural PSC denoiser

+++ adaprobe/psc_denoiser.py
@@ -11,7 +11,7 @@ print('CUDA device: ', torch.cuda.get_device_name())
 class NeuralDenoiser():
 	def __init__(self, n_layers=3, kernel_size=99, padding=49, stride=1):
 		self.denoiser = DenoisingNetwork(n_layers=n_layers, kernel_size=kernel_size,
-			padding=padding, stride=stride)
+			padding=padding, stride=stride).double()
 
 	def __call__(self, traces):
 		# Run denoiser over PSC trace batch
